link QuantState in F.
